Sexto_Cuatrimestre/Blockchain/pyproject.toml
[tool.poetry]
name = "blockchain"
version = "0.1.0"
description = ""
authors = ["Josefina Martinez <josemartinezosti@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.1"
ecdsa = "^0.19.0"
mnemonic = "^0.21"
cryptos = {git = "https://github.com/primal100/pybitcointools"}
eth-utils = "^5.0.0"
eth-keys = "^0.5.1"
eth-hash = {extras = ["pycryptodome"], version = "^0.7.0"}
web3 = "^7.3.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
